requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
albumentations==1.3.0
altair==4.2.0
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==21.4b2
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.8
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
# -f https://dl.fbaipublicfiles.com/detectron2/wheels/cu111/torch1.9/index.html
# detectron2==0.6
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
dpath==2.1.4
dulwich==0.21.3
dvc==2.47.2
dvc-data==0.42.3
dvc-http==2.30.2
dvc-objects==0.21.1
dvc-render==0.2.0
dvc-studio-client==0.6.0
dvc-task==0.2.0
entrypoints==0.4
executing==0.9.0
fastjsonschema==2.16.1
filelock==3.9.0
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.34.4
frozenlist==1.3.3
fsspec==2023.3.0
funcy==1.18
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grandalf==0.8
grpcio==1.51.3
hydra-core==1.3.2
idna==3.3
imageio==2.26.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
iopath==0.1.9
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
iterative-telemetry==0.0.7
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
kombu==5.2.4
lazy_loader==0.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
multidict==6.0.4
mypy-extensions==1.0.0
nanotime==0.5.2
natsort==8.3.1
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.0
notebook==6.4.12
numpy==1.23.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
orjson==3.8.7
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
# pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
portalocker==2.7.0
progress==1.6
progressbar==2.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pydeck==0.7.1
pydot==1.4.2
pygit2==1.11.1
Pygments==2.12.0
pygtrie==2.5.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.0
qudida==0.0.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.8.1
scmrepo==0.1.15
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
split-folders==0.5.1
sqltrie==0.2.0
stack-data==0.3.0
streamlit==1.11.0
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2023.2.28
tinycss2==1.1.1
toml==0.10.2
tomlkit==0.11.6
toolz==0.12.0
-f https://download.pytorch.org/whl/nightly/cpu
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
vine==5.0.0
voluptuous==0.13.1
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==3.6.1
yacs==0.1.8
yarl==1.8.2
zc.lockfile==3.0.post1
zipp==3.8.1